Only clean pid file if it's still around

+++ server/app.rb
@@ -74,7 +74,8 @@ module PrintMe
         Dir['tmp/*.pid'].each do |pid_file|
           pid = File.open(PID_FILE, 'r') { |f| f.read }.to_i
           Process.kill("SIGINT", pid)
-          File.delete(pid_file)
+          # Ensure all pids are cleaned up
+          File.delete(pid_file) if File.exist?(pid_file)
         end
 
         status 200
